backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
yfinance>=0.2.32
requests>=2.31.0
scikit-learn>=1.5.0
pandas>=2.1.3
numpy>=1.26.0
matplotlib>=3.8.0
alpha-vantage>=2.3.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
textblob>=0.17.1
newsapi-python>=0.2.7
google-generativeai>=0.3.0
